Move sflow plugin to inputs directory
